move parse_month/date/number of days to validation
