+++ metrics_utility/management/validation.py
@@ -1,10 +1,13 @@
 import logging
 import os
+import re
 
-from datetime import datetime
+from datetime import datetime, timedelta, timezone
 
-from metrics_utility.automation_controller_billing.helpers import parse_date_param, parse_month, parse_number_of_days
-from metrics_utility.exceptions import BadParameter, MissingRequiredEnvVar, MissingRequiredParameter, UnparsableParameter
+from dateutil import parser
+from dateutil.relativedelta import relativedelta
+
+from metrics_utility.exceptions import BadParameter, DateFormatError, MissingRequiredEnvVar, MissingRequiredParameter, UnparsableParameter
 
 
 # Constants for valid values
@@ -410,3 +413,84 @@ def validate_gather_params(options, HelpText):
             logger.warning('The date for --until is in the future.')
 
     return since, until
+
+
+# should also suport quarters, start/end of month ("4mo_ago_beginning" vs "1 mo ago end"), but doesn't yet
+def parse_date_param(value, help=''):
+    if not value:
+        return None
+
+    value = value.strip().lower()
+    if value.isdigit():
+        raise UnparsableParameter(f'Bare numbers are not valid ({help})')
+
+    now = datetime.now()
+    parsed_date = None
+
+    # N days ago, start of day
+    match = re.fullmatch(r'(\d+)\s*_*(d|da|day|days)(\s*_*ago)?', value)
+    if match:
+        days_ago = int(match.group(1))
+        parsed_date = (now - timedelta(days=days_ago - 1)).replace(hour=0, minute=0, second=0, microsecond=0)
+
+    # N months ago, start of day
+    match = re.fullmatch(r'(\d+)\s*_*(mo|mon|mont|month|months)(\s*_*ago)?', value)
+    if match:
+        months_ago = int(match.group(1))
+        parsed_date = (now - relativedelta(months=months_ago)).replace(hour=0, minute=0, second=0, microsecond=0)
+
+    # N minutes ago
+    match = re.fullmatch(r'(\d+)\s*_*(m|mi|min|minu|minut|minute|minutes)(\s*_*ago)?', value)
+    if match:
+        minutes_ago = int(match.group(1))
+        parsed_date = now - timedelta(minutes=minutes_ago)
+
+    # actual date
+    if not parsed_date:
+        parsed_date = parser.parse(value)
+
+    # Add default UTC timezone
+    if parsed_date and parsed_date.tzinfo is None:
+        parsed_date = parsed_date.replace(tzinfo=timezone.utc)
+
+    return parsed_date
+
+
+def parse_number_of_days(value, help=''):
+    if not value:
+        return None
+
+    value = value.strip().lower()
+    if value.isdigit():
+        raise UnparsableParameter(f'Bare numbers are not valid ({help})')
+
+    # N days ago
+    match = re.fullmatch(r'(\d+)\s*_*(d|da|day|days)', value)
+    if match:
+        return int(match.group(1))
+
+    # N months ago - using 30 days per month
+    match = re.fullmatch(r'(\d+)\s*_*(m|mo|mon|mont|month|months)', value)
+    if match:
+        return int(match.group(1)) * 30
+
+    raise UnparsableParameter(f"Can't parse parameter value {value} ({help})")
+
+
+def parse_month(month):
+    """Process month argument"""
+    if month is not None:
+        try:
+            date = datetime.strptime(f'{month}', '%Y-%m')
+        except ValueError:
+            raise DateFormatError('Invalid --month format. Supported date format: YYYY-MM')
+    else:
+        """Return last month if no month was passed"""
+        beginning_of_the_month = datetime.today().replace(day=1)
+        beginning_of_the_previous_month = beginning_of_the_month - relativedelta(months=1)
+        date = beginning_of_the_previous_month
+        y = date.strftime('%Y')
+        m = date.strftime('%m')
+        month = f'{y}-{m}'
+
+    return month, date, date + relativedelta(months=1)
